Refactor markdown / keys loaders and actions
