+++ apps/website/app/components/translation-key/KeyAction.ts
@@ -1,4 +1,5 @@
 export enum KeyAction {
   Create = "createKey",
   Edit = "editKey",
+  Duplicate = "duplicate",
 }
