+++ apps/website/app/routes.ts
@@ -67,8 +67,8 @@ export default [
           "routes/orgs.$orgSlug.projects.$projectSlug.translations/index.tsx",
         ),
         route(
-          "translations/markdown",
-          "routes/orgs.$orgSlug.projects.$projectSlug.translations.markdown/index.tsx",
+          "translations/files",
+          "routes/orgs.$orgSlug.projects.$projectSlug.translations.files/index.tsx",
         ),
         route(
           "settings",
