+++ apps/website/app/routes/orgs.$orgSlug.projects.$projectSlug.translations/constants.ts
@@ -1,0 +1,6 @@
+export const TRANSLATIONS_LIMIT = 50;
+export enum DocumentMode {
+  Empty = "empty",
+  TranslationKeys = "translationKeys",
+  Document = "document",
+}
