+++ apps/website/app/routes/orgs.$orgSlug.projects.$projectSlug.translations/index.tsx
@@ -1,94 +1,50 @@
+import { Heading, VStack, Button, Box, Text } from "@chakra-ui/react";
 import {
-  Heading,
-  VStack,
-  Button,
-  Box,
-  Text,
-  Stack,
-  Tabs,
-  HStack,
-  IconButton,
-  Badge,
-  Code,
-} from "@chakra-ui/react";
-import {
-  Link,
-  useOutletContext,
   redirect,
-  useActionData,
-  useNavigation,
   useNavigate,
+  useOutletContext,
   useSearchParams,
 } from "react-router";
 import { useTranslation } from "react-i18next";
-import { LuPlus, LuPencil } from "react-icons/lu";
-import { useState, useEffect, useCallback } from "react";
-import {
-  FORMAT_LABELS,
-  SupportedFormat,
-  isSupportedFormat,
-  isDocumentFormat,
-} from "@transi-store/common";
+import { LuPlus } from "react-icons/lu";
+import { isDocumentFormat } from "@transi-store/common";
 import type { Route } from "./+types/index";
 import { userContext } from "~/middleware/auth";
 import { requireOrganizationMembership } from "~/lib/organizations.server";
-import { TranslationKeyDrawer } from "~/components/translation-key";
 import { getProjectBySlug } from "~/lib/projects.server";
-import {
-  getTranslationKeys,
-  duplicateTranslationKey,
-  createTranslationKey,
-  getTranslationKeyByName,
-} from "~/lib/translation-keys.server";
-import {
-  getProjectFiles,
-  updateProjectFile,
-  createProjectFile,
-  deleteProjectFile,
-  DuplicateFilePathError,
-} from "~/lib/project-files.server";
-import { validateOutputPath } from "~/lib/path-utils";
-import { TranslationsSearchBar } from "./TranslationsSearchBar";
-import { TranslationsTable } from "./TranslationsTable";
-import { TranslationsPagination } from "./TranslationsPagination";
-import {
-  TranslationKeyModal,
-  TRANSLATIONS_KEY_MODEL_MODE,
-} from "./TranslationKeyModal";
-import { FileEditModal } from "./FileEditModal";
+import { getProjectFiles } from "~/lib/project-files.server";
 import { getInstance } from "~/middleware/i18next";
-import { getKeyUrl, getTranslationsUrl } from "~/lib/routes-helpers";
-import { TranslationKeysSort } from "~/lib/sort/keySort";
+import {
+  getTranslationsFilesUrl,
+  getTranslationsUrl,
+} from "~/lib/routes-helpers";
 import { createProjectNotFoundResponse } from "~/errors/response-errors/ProjectNotFoundResponse";
-import { FileAction } from "./FileAction";
-import { KeyAction } from "~/components/translation-key/KeyAction";
-
-const LIMIT = 50;
-
-export function resolveSort(
-  sortParam: string | null,
-  hasSearch: boolean,
-): TranslationKeysSort {
-  const sort = Object.values(TranslationKeysSort).includes(
-    sortParam as TranslationKeysSort,
-  )
-    ? (sortParam as TranslationKeysSort)
-    : undefined;
-  if (!hasSearch && sort === TranslationKeysSort.RELEVANCE) {
-    return TranslationKeysSort.ALPHABETICAL;
-  }
-  return (
-    sort ??
-    (hasSearch
-      ? TranslationKeysSort.RELEVANCE
-      : TranslationKeysSort.ALPHABETICAL)
-  );
-}
+import { ProjectFileTabs } from "~/components/project-files/ProjectFileTabs";
+import { FileEditModal } from "./FileEditModal";
+import { TranslationKeysView } from "./TranslationKeysView";
+import { DocumentTranslationsView } from "./DocumentTranslationsView";
+import {
+  translationKeysLoader,
+  resolveSort,
+} from "./loadTranslationKeys.server";
+import { loadDocumentTranslations } from "./loadDocumentTranslations.server";
+import {
+  isMarkdownAction,
+  runMarkdownAction,
+} from "./runMarkdownAction.server";
+import { isKeyAction, runKeyAction } from "./runKeyAction.server";
+import type { ProjectFile } from "../../../drizzle/schema";
+import { DocumentMode } from "./constants";
 
 type ContextType = {
   organization: { id: string; slug: string; name: string };
   project: { id: string; slug: string; name: string };
   languages: Array<{ id: string; locale: string; isDefault: boolean }>;
+};
+
+type EmptyLoaderData = {
+  mode: DocumentMode.Empty;
+  projectFiles: Array<ProjectFile>;
 };
 
 export async function loader({ request, params, context }: Route.LoaderArgs) {
@@ -109,20 +65,14 @@ export async function loader({ request, params, context }: Route.LoaderArgs) {
   const highlight = url.searchParams.get("highlight") || undefined;
   const sort = resolveSort(url.searchParams.get("sort"), Boolean(search));
   const page = parseInt(url.searchParams.get("page") || "1", 10);
-  const offset = (page - 1) * LIMIT;
 
   const projectFiles = await getProjectFiles(project.id);
 
   if (projectFiles.length === 0) {
     return {
-      keys: { data: [], count: 0 },
+      mode: DocumentMode.Empty,
       projectFiles,
-      selectedFileId: null,
-      search,
-      highlight,
-      page,
-      sort,
-    };
+    } satisfies EmptyLoaderData;
   }
 
   const fileIdParam = url.searchParams.get("fileId");
@@ -142,31 +92,23 @@ export async function loader({ request, params, context }: Route.LoaderArgs) {
     );
   }
 
-  // Files in a document format (markdown / mdx) use a different UI: redirect
-  // to the dedicated full-width side-by-side editor.
   if (isDocumentFormat(selectedFile.format)) {
-    throw redirect(
-      `/orgs/${params.orgSlug}/projects/${params.projectSlug}/translations/markdown?fileId=${selectedFile.id}`,
-    );
+    return loadDocumentTranslations({
+      projectId: project.id,
+      projectFiles,
+      selectedFile,
+    });
   }
 
-  const keys = await getTranslationKeys(project.id, {
-    search,
-    limit: LIMIT,
-    offset,
-    sort,
-    fileId: selectedFile.id,
-  });
-
-  return {
-    keys,
+  return translationKeysLoader({
+    projectId: project.id,
     projectFiles,
     selectedFileId: selectedFile.id,
     search,
     highlight,
     page,
     sort,
-  };
+  });
 }
 
 export async function action({ request, params, context }: Route.ActionArgs) {
@@ -186,215 +128,26 @@ export async function action({ request, params, context }: Route.ActionArgs) {
   const formData = await request.formData();
   const action = formData.get("_action");
 
-  if (action === "duplicate") {
-    const keyId = formData.get("keyId");
-
-    if (!keyId || typeof keyId !== "string") {
-      throw new Response("Key ID is required", { status: 400 });
-    }
-
-    const parsedKeyId = parseInt(keyId, 10);
-
-    if (isNaN(parsedKeyId)) {
-      throw new Response("Invalid Key ID", { status: 400 });
-    }
-
-    const newKeyId = await duplicateTranslationKey(parsedKeyId);
-
-    return redirect(getKeyUrl(params.orgSlug, params.projectSlug, newKeyId));
-  }
-
-  if (action === KeyAction.Create) {
-    const keyName = formData.get("keyName");
-    const description = formData.get("description");
-    const fileIdRaw = formData.get("fileId");
-
-    if (!keyName || typeof keyName !== "string") {
-      return {
-        error: i18next.t("keys.new.errors.nameRequired"),
-        action: KeyAction.Create,
-      };
-    }
-
-    if (!fileIdRaw || typeof fileIdRaw !== "string") {
-      return {
-        error: i18next.t("files.errors.missingFileId"),
-        action: KeyAction.Create,
-      };
-    }
-    const fileId = parseInt(fileIdRaw, 10);
-    if (isNaN(fileId)) {
-      return {
-        error: i18next.t("files.errors.invalidFileId", { fileId: fileIdRaw }),
-        action: KeyAction.Create,
-      };
-    }
-
-    // Vérifier que la clé n'existe pas déjà
-    const existing = await getTranslationKeyByName(project.id, keyName, fileId);
-    if (existing) {
-      return {
-        error: i18next.t("keys.new.errors.alreadyExists", { keyName }),
-        action: KeyAction.Create,
-      };
-    }
-
-    // Créer la clé
-    const newKeyId = await createTranslationKey({
+  if (isMarkdownAction(action)) {
+    return runMarkdownAction({
+      action,
+      formData,
       projectId: project.id,
-      keyName,
-      description:
-        description && typeof description === "string"
-          ? description
-          : undefined,
-      fileId,
+      organizationId: organization.id,
+      projectSlug: params.projectSlug,
+      i18next,
     });
-
-    // Retourner le succès avec l'ID et le nom de la clé
-    return {
-      success: true,
-      keyId: newKeyId,
-      keyName,
-      search: keyName,
-      action: KeyAction.Create,
-    };
   }
 
-  if (action === FileAction.Create) {
-    const filePath = formData.get("filePath");
-    const fileFormat = formData.get("fileFormat");
-
-    if (!filePath || typeof filePath !== "string") {
-      return {
-        error: i18next.t("files.errors.pathRequired"),
-        action: FileAction.Create,
-      };
-    }
-    const pathError = validateOutputPath(filePath, i18next.t);
-    if (pathError) {
-      return { error: pathError, action: FileAction.Create };
-    }
-
-    if (
-      !fileFormat ||
-      typeof fileFormat !== "string" ||
-      !isSupportedFormat(fileFormat)
-    ) {
-      return {
-        error: i18next.t("files.errors.invalidFormat"),
-        action: FileAction.Create,
-      };
-    }
-
-    try {
-      const created = await createProjectFile(project.id, {
-        filePath,
-        format: fileFormat as SupportedFormat,
-      });
-
-      return redirect(
-        getTranslationsUrl(params.orgSlug, params.projectSlug, {
-          fileId: created.id,
-        }),
-      );
-    } catch (error) {
-      if (error instanceof DuplicateFilePathError) {
-        return {
-          error: i18next.t("files.errors.duplicatePath", { filePath }),
-          action: FileAction.Create,
-        };
-      }
-      throw error;
-    }
-  }
-
-  if (action === FileAction.Edit) {
-    const fileId = formData.get("fileId");
-    const filePath = formData.get("filePath");
-    const fileFormat = formData.get("fileFormat");
-
-    if (!fileId || typeof fileId !== "string") {
-      return {
-        error: i18next.t("files.errors.missingFileId"),
-        action: FileAction.Edit,
-      };
-    }
-    const parsedFileId = parseInt(fileId, 10);
-    if (isNaN(parsedFileId)) {
-      return {
-        error: i18next.t("files.errors.invalidFileId", { fileId }),
-        action: FileAction.Edit,
-      };
-    }
-
-    if (!filePath || typeof filePath !== "string") {
-      return {
-        error: i18next.t("files.errors.pathRequired"),
-        action: FileAction.Edit,
-      };
-    }
-    const pathError = validateOutputPath(filePath, i18next.t);
-    if (pathError) {
-      return { error: pathError, action: FileAction.Edit };
-    }
-
-    if (
-      !fileFormat ||
-      typeof fileFormat !== "string" ||
-      !isSupportedFormat(fileFormat)
-    ) {
-      return {
-        error: i18next.t("files.errors.invalidFormat"),
-        action: FileAction.Edit,
-      };
-    }
-
-    try {
-      await updateProjectFile(project.id, parsedFileId, {
-        filePath,
-        format: fileFormat as SupportedFormat,
-      });
-    } catch (error) {
-      if (error instanceof DuplicateFilePathError) {
-        return {
-          error: i18next.t("files.errors.duplicatePath", { filePath }),
-          action: FileAction.Edit,
-        };
-      }
-      throw error;
-    }
-
-    const url = new URL(request.url);
-    return redirect(
-      getTranslationsUrl(params.orgSlug, params.projectSlug, {
-        fileId: url.searchParams.get("fileId") ?? undefined,
-        search: url.searchParams.get("search") || undefined,
-        sort: url.searchParams.get("sort") || undefined,
-        highlight: url.searchParams.get("highlight") || undefined,
-      }),
-    );
-  }
-
-  if (action === FileAction.Delete) {
-    const fileId = formData.get("fileId");
-
-    if (!fileId || typeof fileId !== "string") {
-      return {
-        error: i18next.t("files.errors.missingFileId"),
-        action: FileAction.Delete,
-      };
-    }
-    const parsedFileId = parseInt(fileId, 10);
-    if (isNaN(parsedFileId)) {
-      return {
-        error: i18next.t("files.errors.invalidFileId", { fileId }),
-        action: FileAction.Delete,
-      };
-    }
-
-    await deleteProjectFile(project.id, parsedFileId);
-
-    return redirect(getTranslationsUrl(params.orgSlug, params.projectSlug));
+  if (isKeyAction(action)) {
+    return runKeyAction({
+      action,
+      formData,
+      projectId: project.id,
+      orgSlug: params.orgSlug,
+      projectSlug: params.projectSlug,
+      i18next,
+    });
   }
 
   throw new Response(i18next.t("keys.errors.unknownAction"), { status: 400 });
@@ -403,23 +156,11 @@ export async function action({ request, params, context }: Route.ActionArgs) {
 export default function ProjectTranslations({
   loaderData,
 }: Route.ComponentProps) {
+  const context = useOutletContext<ContextType>();
   const { t } = useTranslation();
-  const {
-    keys: { data, count },
-    projectFiles,
-    selectedFileId,
-    search,
-    highlight,
-    page,
-    sort,
-  } = loaderData;
-  const { organization, project, languages } = useOutletContext<ContextType>();
-  const actionData = useActionData<typeof action>();
-  const navigation = useNavigation();
   const navigate = useNavigate();
-  const isSubmitting = navigation.state === "submitting";
-
   const [searchParams] = useSearchParams();
+
   const modal = searchParams.get("modal");
   const isCreateFileModalOpen = modal === "create-file";
   const editingFileId = (() => {
@@ -427,9 +168,10 @@ export default function ProjectTranslations({
     const id = parseInt(modal.slice("edit-file-".length), 10);
     return isNaN(id) ? null : id;
   })();
-  const editingFile = projectFiles.find((f) => f.id === editingFileId) ?? null;
-
-  const [isCreateKeyModalOpen, setIsCreateKeyModalOpen] = useState(false);
+  const editingFile =
+    loaderData.mode === DocumentMode.Empty
+      ? null
+      : (loaderData.projectFiles.find((f) => f.id === editingFileId) ?? null);
 
   function openFileModal(type: string) {
     const next = new URLSearchParams(searchParams);
@@ -443,270 +185,109 @@ export default function ProjectTranslations({
     navigate(`?${next.toString()}`);
   }
 
-  // Drawer state for inline editing
-  const [drawerKeyId, setDrawerKeyId] = useState<number | null>(null);
+  const filesActionUrl = getTranslationsFilesUrl(
+    context.organization.slug,
+    context.project.slug,
+    loaderData.mode === DocumentMode.TranslationKeys
+      ? {
+          search: loaderData.search,
+          sort: loaderData.sort,
+          highlight: loaderData.highlight,
+          fileId: loaderData.selectedFileId,
+        }
+      : loaderData.mode === DocumentMode.Document
+        ? { fileId: loaderData.selectedFileId }
+        : undefined,
+  );
 
-  const handleEditInDrawer = useCallback((keyId: number) => {
-    setDrawerKeyId(keyId);
-  }, []);
-
-  const handleDrawerClosed = useCallback(() => {
-    setDrawerKeyId(null);
-  }, []);
-
-  const totalLanguages = languages.length;
-
-  // Build redirect URL with current search params
-  const currentUrl = getTranslationsUrl(organization.slug, project.slug, {
-    search,
-    sort,
-    highlight,
-    fileId: selectedFileId,
-  });
-
-  // Close modal and navigate after successful creation
-  useEffect(() => {
-    if (
-      actionData?.success &&
-      actionData.action === KeyAction.Create &&
-      navigation.state === "idle"
-    ) {
-      // eslint-disable-next-line react-hooks/set-state-in-effect
-      setIsCreateKeyModalOpen(false);
-
-      // Navigate to filter by the newly created key
-      navigate(
-        getTranslationsUrl(organization.slug, project.slug, {
-          fileId: selectedFileId,
-          sort: TranslationKeysSort.CREATED_AT,
-          highlight: highlight
-            ? `${highlight},${actionData.keyName}`
-            : actionData.keyName,
-        }),
-      );
-
-      // Open the drawer for the newly created key
-      if (actionData.keyId) {
-        setDrawerKeyId(actionData.keyId);
-      }
-    }
-  }, [
-    actionData,
-    navigation.state,
-    organization.slug,
-    project.slug,
-    navigate,
-    highlight,
-    selectedFileId,
-  ]);
+  const isFileModalOpen = isCreateFileModalOpen || editingFile !== null;
+  const selectedFile =
+    loaderData.mode === DocumentMode.Empty
+      ? undefined
+      : loaderData.projectFiles.find((f) => f.id === loaderData.selectedFileId);
 
   return (
     <VStack gap={6} align="stretch">
-      <Stack
-        direction={{ base: "column", sm: "row" }}
-        justify="space-between"
-        align={{ base: "stretch", sm: "center" }}
-        gap={{ base: 3, sm: 0 }}
-      >
-        <Box>
-          <Heading as="h2" size="lg">
-            {t("translations.title")}
-          </Heading>
-          <Text color="gray" mt={2}>
-            {t("translations.count", { count })}
-          </Text>
-        </Box>
-        {languages.length > 0 && projectFiles.length > 0 && (
-          <Button
-            colorPalette="accent"
-            onClick={() => setIsCreateKeyModalOpen(true)}
-            width={{ base: "full", sm: "auto" }}
-          >
-            <LuPlus /> {t("translations.newKey")}
-          </Button>
-        )}
-      </Stack>
+      {loaderData.mode === DocumentMode.Empty ? (
+        <>
+          <Box>
+            <Heading as="h2" size="lg">
+              {t("translations.title")}
+            </Heading>
+          </Box>
 
-      {projectFiles.length === 0 ? (
-        <Box
-          p={10}
-          textAlign="center"
-          borderWidth={1}
-          borderRadius="lg"
-          bg="bg.subtle"
-        >
-          <Text color="fg.muted" mb={4}>
-            {t("files.noFiles")}
-          </Text>
-          <Button
-            colorPalette="accent"
-            onClick={() => openFileModal("create-file")}
+          <Box
+            p={10}
+            textAlign="center"
+            borderWidth={1}
+            borderRadius="lg"
+            bg="bg.subtle"
           >
-            <LuPlus /> {t("files.addFile")}
-          </Button>
-        </Box>
-      ) : (
-        <Tabs.Root
-          value={selectedFileId !== null ? String(selectedFileId) : undefined}
-          variant="line"
-          size="sm"
-        >
-          <HStack align="center" gap={2} wrap="wrap">
-            <Tabs.List flex="1" minW={0}>
-              {projectFiles.map((file) => (
-                <HStack key={file.id} gap={0} align="center">
-                  <Tabs.Trigger
-                    value={String(file.id)}
-                    cursor="pointer"
-                    onClick={() => {
-                      if (file.id === selectedFileId) return;
-                      navigate(
-                        getTranslationsUrl(organization.slug, project.slug, {
-                          fileId: file.id,
-                          search,
-                          sort,
-                          highlight,
-                        }),
-                      );
-                    }}
-                  >
-                    <Code fontSize="xs">{file.filePath}</Code>
-                    <Badge size="xs" ml={2}>
-                      {FORMAT_LABELS[file.format as SupportedFormat] ??
-                        file.format}
-                    </Badge>
-                  </Tabs.Trigger>
-                  <IconButton
-                    aria-label={t("files.editFile")}
-                    size="xs"
-                    variant="ghost"
-                    onClick={() => openFileModal(`edit-file-${file.id}`)}
-                  >
-                    <LuPencil />
-                  </IconButton>
-                </HStack>
-              ))}
-            </Tabs.List>
-            <IconButton
-              aria-label={t("files.addFile")}
-              size="xs"
-              variant="ghost"
+            <Text color="fg.muted" mb={4}>
+              {t("files.noFiles")}
+            </Text>
+            <Button
+              colorPalette="accent"
               onClick={() => openFileModal("create-file")}
             >
-              <LuPlus />
-            </IconButton>
-          </HStack>
-        </Tabs.Root>
-      )}
-
-      {projectFiles.length > 0 && (
-        <TranslationsSearchBar
-          search={search}
-          sort={sort}
-          organizationSlug={organization.slug}
-          projectSlug={project.slug}
-          fileId={selectedFileId ?? undefined}
-        />
-      )}
-
-      {projectFiles.length === 0 ? null : languages.length === 0 ? (
-        <Box p={10} textAlign="center" borderWidth={1} borderRadius="lg">
-          <Text color="fg.muted" mb={4}>
-            {t("translations.noLanguages")}
-          </Text>
-          <Button asChild colorPalette="brand">
-            <Link
-              to={`/orgs/${organization.slug}/projects/${project.slug}/settings`}
-            >
-              {t("translations.manageLanguages")}
-            </Link>
-          </Button>
-        </Box>
-      ) : data.length === 0 ? (
-        <Box p={8} textAlign="center" bg="bg.subtle" borderRadius="md">
-          <Text color="fg.muted">
-            {search
-              ? t("translations.noResultsForSearch")
-              : t("translations.noKeysEmpty")}
-          </Text>
-        </Box>
+              <LuPlus /> {t("files.addFile")}
+            </Button>
+          </Box>
+        </>
       ) : (
         <>
-          <TranslationsTable
-            data={data}
-            search={search}
-            totalLanguages={totalLanguages}
-            organizationSlug={organization.slug}
-            projectSlug={project.slug}
-            currentUrl={currentUrl}
-            onEditInDrawer={handleEditInDrawer}
+          <ProjectFileTabs
+            files={loaderData.projectFiles}
+            selectedFileId={loaderData.selectedFileId}
+            onFileClick={(file) => {
+              if (file.id === loaderData.selectedFileId) return;
+              navigate(
+                getTranslationsUrl(
+                  context.organization.slug,
+                  context.project.slug,
+                  {
+                    fileId: file.id,
+                    search:
+                      loaderData.mode === DocumentMode.TranslationKeys
+                        ? loaderData.search
+                        : undefined,
+                    sort:
+                      loaderData.mode === DocumentMode.TranslationKeys
+                        ? loaderData.sort
+                        : undefined,
+                    highlight:
+                      loaderData.mode === DocumentMode.TranslationKeys
+                        ? loaderData.highlight
+                        : undefined,
+                  },
+                ),
+              );
+            }}
+            onEditFile={(file) => openFileModal(`edit-file-${file.id}`)}
+            onAddFile={() => openFileModal("create-file")}
           />
 
-          <TranslationsPagination
-            count={count}
-            pageSize={LIMIT}
-            currentPage={page}
-            search={search}
-            sort={sort}
-            organizationSlug={organization.slug}
-            projectSlug={project.slug}
-            fileId={selectedFileId ?? undefined}
-          />
+          {loaderData.mode === DocumentMode.Document && selectedFile ? (
+            <DocumentTranslationsView
+              data={loaderData}
+              context={context}
+              selectedFile={selectedFile}
+            />
+          ) : loaderData.mode === DocumentMode.TranslationKeys ? (
+            <TranslationKeysView data={loaderData} context={context} />
+          ) : null}
         </>
       )}
 
-      {/* Drawer for inline key editing */}
-      {drawerKeyId !== null && (
-        <TranslationKeyDrawer
-          keyId={drawerKeyId}
-          organizationSlug={organization.slug}
-          projectSlug={project.slug}
-          redirectUrl={currentUrl}
-          onClosed={handleDrawerClosed}
-        />
-      )}
-
-      {/* Modale de création de clé */}
-      {selectedFileId !== null && (
-        <TranslationKeyModal
-          isOpen={isCreateKeyModalOpen}
-          onOpenChange={setIsCreateKeyModalOpen}
-          mode={TRANSLATIONS_KEY_MODEL_MODE.CREATE}
-          error={
-            actionData?.error && actionData.action === KeyAction.Create
-              ? actionData.error
-              : undefined
-          }
-          isSubmitting={isSubmitting}
-          fileId={selectedFileId}
-        />
-      )}
-
-      {editingFile && (
-        <FileEditModal
-          isOpen={true}
-          onOpenChange={(open) => {
-            if (!open) closeFileModal();
-          }}
-          file={editingFile}
-          error={
-            actionData?.error && actionData.action === FileAction.Edit
-              ? actionData.error
-              : undefined
-          }
-        />
-      )}
-
       <FileEditModal
-        isOpen={isCreateFileModalOpen}
+        isOpen={isFileModalOpen}
         onOpenChange={(open) => {
-          if (!open) closeFileModal();
+          if (!open) {
+            closeFileModal();
+          }
         }}
-        error={
-          actionData?.error && actionData.action === FileAction.Create
-            ? actionData.error
-            : undefined
-        }
+        file={editingFile}
+        actionUrl={filesActionUrl}
       />
     </VStack>
   );
